+++ docs/Whats-New.md
@@ -54,6 +54,7 @@ New:
 - Automatic Passenger Deletion (by FS-21)
 - Script Action 74 to 81 and 84 to 91 for new AI attacks (by FS-21)
 - Script Actions 82 & 83 for modifying AI Trigger Current Weight (by FS-21)
+- Script Action 92 for waiting & repeat the same new AI attack if no target was found (by FS-21)
 - Script Action 112 to regroup temporarily around the Team Leader (by FS-21)
 
 Vanilla fixes: